Fixing associations on text change (#20)(#24)
When the user loaded a new text using the 'Load new text' button on
Project properties dialog iBiblia retained the original value of the
associations field. This always resulted in corrupted associations if
the new text was different.

Now iBiblia asks the user if he wants to discard all associations or try
to keep them. If the user wants to keep them, iBiblia uses the 'replace
text' feature of the TVersiculo class, which tries to keep associations
for equal words. This algorithm still oversees some associations, but
now we never should see the 'corrupt associations' error anymore.

The 'Clear text' button also did not clear the associtions.

Another bug fixed was with the 'Discard verse associations' button: it
clear the TVersiculo's only but kept the database associations until the
user changed verses. This seemed to lead to some associations from an
old associated verse crawling to an unrelated verse.