Re-enabling Ctrl+MouseWheel for zooming on popup
The mouse-wheel zooming feature is not available by default on the
TRichMemo component.

There still a bug: the first time the popup is shown the saved zoom is
not applied. This is a TRichMemo bug. I'll try and circunvent it later.

Also made the popup windows more like a THint, hiding completely the
caption and borders.