ChapterView: Avoiding wrong-size notes
Working around what seems to be a bug in TKMemo.ClientRect, which
returns 5x5 the first time even after we have set its text; the other
problem is that it returns an slightly smaller rect, resulting in
scrollbars. It turns out that if we call it again it gives back right
results.