Better handling of synchronization error
Now the synchronization with theWord works if the user opens a new
theWord instance with a different language.

The only situation it cannot recover so far is when the user changes
theWord's language and do not restart it. The problem is that theWord
does not update the config file right away, but only when theWord is
closed. A message dialog is now display explaining this situation.