True interlinear mode
Implementation of a true interlinear mode in chapter view. The user can
choose between the old INTRAlinear mode, the one used in theWord, and
the new INTERlinear mode.

The user can apply custom verse rules independently for both of them.
Added also the option to disable a verse rule and the option to apply
them to the exported interlinear Bible module. With this it's possible
to work with an interlinear source text and have a correct interlinear
output, stripping out the original glosses using the verse rules.

Added code to detect verse rules errors and give the user the chance to
fix them.

Added a button in the interlinear verse rules window to copy the
chapter's raw content to the clipboard. This can be very useful to try
the verse rules in an editor before adding them to iBiblia.

Also, some options there were save to the iBiblia.ini file are now saved
to the project file itself. Those are the current chapter view tab,
paragraph mode and the new interlinear mode option.